[Refactor] Move JVM metrics into Udf
