+++ be/src/udf/CMakeLists.txt
@@ -21,6 +21,7 @@ set(EXECUTABLE_OUTPUT_PATH "${BUILD_DIR}/src/udf")
 
 set(UDF_FILES
   java/utils.cpp
+  java/jvm_metrics.cpp
   java/java_udf.cpp
   java/java_data_converter.cpp
   java/java_native_method.cpp
