+++ be/test/CMakeLists.txt
@@ -295,7 +295,7 @@ set(EXEC_FILES
         ./udf/java/java_native_method_test.cpp
         ./udf/java/java_data_converter_test.cpp
         ./udf/java/java_udf_test.cpp
-        ./util/jvm_metrics_test.cpp
+        ./udf/java/jvm_metrics_test.cpp
         ./util/network_util_test.cpp
         )
